chore: moving to new route structure
