+++ app/routes/[sitemap.xml].tsx
@@ -1,8 +1,8 @@
 import type { LoaderFunction } from "@remix-run/node";
 
 import { fetchFromGraphCMS } from "~/utils/graphcms";
-import type { Post } from "~/routes/api/blog";
-import type { Portfolio } from "~/routes/api/portfolio";
+import type { Post } from "~/routes/api.blog";
+import type { Portfolio } from "~/routes/api.portfolio";
 import { BASE_URL } from "~/config/settings.server";
 import { SITE_UPDATED } from "~/config/constants";
 import { getSitemap } from "~/queries/getSitemap";
