+++ app/routes/api.blog.$slug.ts
