+++ app/routes/api.portfolio.$slug.ts
