+++ app/routes/blog.tsx
@@ -4,8 +4,8 @@ import type { MetaFunction } from "@remix-run/node";
 import { BlogPreview } from "~/components/BlogPreview";
 import { AppHero } from "~/components/AppHero";
 import { SITE_AUTHOR, SITE_TITLE } from "~/config/constants";
-import { loader } from "~/routes/api/blog";
-import type { LoaderData } from "~/routes/api/blog";
+import { loader } from "~/routes/api.blog";
+import type { LoaderData } from "~/routes/api.blog";
 import { BlogFeatured } from "~/components/BlogFeatured";
 import { BlogUpcoming } from "~/components/BlogUpcoming";
 import { getMetaData } from "~/metadata";
