+++ app/routes/blog_.$slug.tsx
@@ -2,8 +2,8 @@ import { useCatch, useLoaderData } from "@remix-run/react";
 import type { LinksFunction, MetaFunction } from "@remix-run/node";
 
 import { AppHero } from "~/components/AppHero";
-import type { LoaderData } from "~/routes/api/blog/$slug";
-import { loader } from "~/routes/api/blog/$slug";
+import type { LoaderData } from "~/routes/api.blog.$slug";
+import { loader } from "~/routes/api.blog.$slug";
 import { AppWysiwyg } from "~/components/AppWysiwyg";
 
 import stylesLines from "prismjs/plugins/line-numbers/prism-line-numbers.css";
