+++ app/routes/portfolio_.$slug.tsx
@@ -2,8 +2,8 @@ import { useCatch, useLoaderData } from "@remix-run/react";
 import type { MetaFunction } from "@remix-run/node";
 
 import { AppHero } from "~/components/AppHero";
-import { loader } from "~/routes/api/portfolio/$slug";
-import type { LoaderData } from "~/routes/api/portfolio/$slug";
+import { loader } from "~/routes/api.portfolio.$slug";
+import type { LoaderData } from "~/routes/api.portfolio.$slug";
 import { getMetaData } from "~/metadata";
 
 export { loader };
