+++ package.json
@@ -20,9 +20,9 @@
   "dependencies": {
     "@fireworks-js/react": "^2.10.2",
     "@graphcms/rich-text-react-renderer": "^0.6.1",
-    "@remix-run/node": "1.13.0",
-    "@remix-run/react": "1.13.0",
-    "@remix-run/vercel": "1.13.0",
+    "@remix-run/node": "^1.14.1",
+    "@remix-run/react": "^1.14.1",
+    "@remix-run/vercel": "^1.14.1",
     "@vercel/analytics": "^0.1.11",
     "classnames": "^2.3.2",
     "date-fns": "^2.29.3",
@@ -35,9 +35,9 @@
     "reading-time": "^1.5.0"
   },
   "devDependencies": {
-    "@remix-run/dev": "1.13.0",
-    "@remix-run/eslint-config": "1.13.0",
-    "@remix-run/serve": "1.13.0",
+    "@remix-run/dev": "^1.14.1",
+    "@remix-run/eslint-config": "^1.14.1",
+    "@remix-run/serve": "^1.14.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
@@ -46,12 +46,12 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "@vitejs/plugin-react": "^3.1.0",
-    "autoprefixer": "^10.4.13",
+    "autoprefixer": "^10.4.14",
     "concurrently": "^7.6.0",
     "cross-env": "^7.0.3",
     "cssnano": "^5.1.15",
-    "eslint": "^8.33.0",
-    "jsdom": "^21.1.0",
+    "eslint": "^8.36.0",
+    "jsdom": "^21.1.1",
     "postcss": "^8.4.21",
     "postcss-cli": "^10.1.0",
     "postcss-import": "^15.1.0",
