+++ remix.config.js
@@ -2,6 +2,9 @@
  * @type {import('@remix-run/dev').AppConfig}
  */
 module.exports = {
+  future: {
+    v2_routeConvention: true
+  },
   serverBuildDirectory: "build",
   serverDependenciesToBundle: [
     "decode-uri-component",
